requirements.txt
PyGObject>=3.42.0
Pillow>=10.0.0
pycairo>=1.20.0
